+++ Projetos/spring_mte/src/main/java/com/mte2023/spring_mte/entities/Address.java
@@ -2,6 +2,8 @@ package com.mte2023.spring_mte.entities;
 
 import java.io.Serializable;
 import java.util.List;
+import java.util.Objects;
+
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -33,7 +35,7 @@ public class Address implements Serializable{
 
     @JsonIgnore
     @OneToMany(mappedBy = "addressId")
-    private List<CollectionOrder> collectionOrderList;
+    private List<User> userList;
 
     public Address() {}
 
@@ -87,12 +89,12 @@ public class Address implements Serializable{
         this.state = state;
     }
 
-    public List<CollectionOrder> getCollectionOrderList() {
-        return collectionOrderList;
+    public List<User> getUserList() {
+        return userList;
     }
 
-    public void setCollectionOrderList(List<CollectionOrder> collectionOrderList) {
-        this.collectionOrderList = collectionOrderList;
+    public void setUserList(List<User> userList) {
+        this.userList = userList;
     }
 
     public String toJson() {
@@ -103,29 +105,18 @@ public class Address implements Serializable{
             throw new RuntimeException(e);
         }
     }
+
     @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((id == null) ? 0 : id.hashCode());
-        return result;
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Address address = (Address) o;
+        return Objects.equals(id, address.id);
     }
 
     @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        Address other = (Address) obj;
-        if (id == null) {
-            if (other.id != null)
-                return false;
-        } else if (!id.equals(other.id))
-            return false;
-        return true;
+    public int hashCode() {
+        return Objects.hash(id);
     }
 
     private static class Builder{
