+++ Projetos/spring_mte/src/main/java/com/mte2023/spring_mte/entities/CollectionOrder.java
@@ -34,18 +34,13 @@ public class CollectionOrder implements Serializable {
 
     @Column(name = "user_id")
     private Integer userID;
-    @Column(name = "address_id")
-    private Integer addressID;
 
     @ManyToOne
     @JoinColumn(name = "user_id", referencedColumnName = "id", insertable = false, updatable = false)
     @JsonIgnore
     private User userId;
 
-    @ManyToOne
-    @JoinColumn(name = "address_id",referencedColumnName = "id", insertable = false, updatable = false)
-    @JsonIgnore
-    private Address addressId;
+
 
  //   private Payment payment;
     public CollectionOrder(){}
@@ -86,22 +81,6 @@ public class CollectionOrder implements Serializable {
     }
     public void setUserId(User userId) {
         this.userId = userId;
-    }
-
-    public Integer getAddressID() {
-        return addressID;
-    }
-
-    public void setAddressID(Integer addressID) {
-        this.addressID = addressID;
-    }
-
-    public Address getAddressId() {
-        return addressId;
-    }
-
-    public void setAddressId(Address addressId) {
-        this.addressId = addressId;
     }
 
     /*
@@ -175,14 +154,7 @@ public class CollectionOrder implements Serializable {
             collectionOrder.userId = userId;
             return this;
         }
-        public Builder withAddressID(Integer addressID){
-            collectionOrder.addressID = addressID;
-            return this;
-        }
-        public Builder withAddressId(Address addressId){
-            collectionOrder.addressId = addressId;
-            return this;
-        }
+
         public CollectionOrder build(){
             return collectionOrder;
         }
